Enable trickling of end-of-candidates through addIceCandidate.
A null parameter in addIceCandidate now represents end-of-candidates
for all media sections.

The definitions of the `failed` and `completed` ICE connection states
have been updated so that they only occur when there are no more
remote candidates.

Also updated the example code that trickled candidates, such that the
null candidate is trickled rather than ignored.