Splitting apart RTCIceConnectionState and RTCIceTransportState.
The fundamental state definitions, example state transitions, and
non-normative state transition diagram now all move to the
RTCIceTransportState section, and are defined in terms of
RTCIceTransports instead of an "ICE Agent".

The RTCIceConnectionState section now contains explicit rules for how
multiple RTCIceTransport states are aggregated, similar to how the
RTCPeerConnectionState is defined. So, situations that were previously
ambiguous (such as "what if one transport is 'new' and the other
is 'connected'?") are now well-defined.