Specify how to handle invalid data channel IDs, or lack of IDs.
Fixes issue #746.

For an ID of 65535, createDataChannel will throw an exception. If
it's determined that an ID is invalid for any other reason (such as
shortage of negotiated SCTP streams), or if no more valid IDs are
available when the user agent tries to pick an ID, the data channel will
fire an "error" event and be closed.

For now, I'm leaving the actual error as a "NetworkError", since I
think we need to separately decide how error events on data channels
work (see issue #846).