Clarify when removeTrack should throw an exception.
In PR #844, the intent was to throw an exception when the sender wasn't created
by the connection on which the method was called. And the assumption was that
senders are never removed, so the algorithm can simply check if it's in the
connection's set of senders.

But something I didn't consider is that rollbacks are able to remove
transceivers (and thus senders). So more clarity is needed.