Moving "request identity assertion" to the synchronous steps.
This simplifies the algorithm a bit, and ignores the case of:

`pc.createOffer()`
`setIdentityProvider(idp)`

Only the identity provider set at the time `createOffer` was called will
be used. And the identity assertion request process will begin as soon
as `createOffer` is called, rather than when the operation is taken off
the queue.