Revise text dealing with with the ICE Agent/User Agent interactions.
Addresses issue #801.

This commit removes most of the text dealing with the ICE Agent from the
RTCPeerConnection section, which was somewhat outdated and incomplete
due to the introduction of `RTCIceTransport`s, additions to the
definitions of the ICE states, etc. Instead, these interactions are now
described in the `RTCIceTransport` section. There are four indications
provided by the ICE Agent:
- Gathering state changed
- Candidate gathered
- Connection state changed
- Selected pair changed

Each of these will queue a task which will update the RTCIceTransport
object, and update the RTCPeerConnection's aggregate connection state
and ICE state if necessary.

Other minor changes in this commit:
- Fix up `setConfiguration` to match JSEP better (for properties related
  to the ICE Agent).
- In `RTCPeerConnection.close()`, set `RTCIceTransport` state to
  "closed" immediately.