Make RTCDataChannel.id nullable and describe when it's set.
Fixes #795.

When a data channel is initially created, and a ID isn't explicitly
provided, it will only be set after the DTLS role is negotiated by an
SDP answer. If a data channel is created after this answer is set, an
ID will be set instantly.

Until the id is set, the attribute will return null.