Switching to new, consistent terminology when talking about exceptions.
Fixes #845.

Throwing exceptions will use the phrase "throw a FooError", which
implies that the current steps will be aborted, making it unnecessary
to clarify that.

Rejecting a promise with an exception will use the phrase "reject p
with a newly created FooError and abort these steps."

Each place where these phrases are used links to the "Terminology"
section, which links to where these terms are defined in WebIDL.