Specify which encoding parameters can be set through sendEncodings.
Fixes #1073.

All read-write parameters and `rid` can be set, but the rest are
ignored, as decided in the April virtual interim. Also added some detail
to `getParameters` description, explaining that it should return the
previously provided `sendEncodings` even if `setRemoteDescription`
hasn't been called yet.

I'm not yet addressing the simulcast error point that arose in the
discussion of the issue. We can address that after we finish fleshing
out how simulcast errors will be handled in the first place.