Specify how data channel priority enum is initialized from priority int
Fixes #1423.

The priority value from the data channel "OPEN" message is a two-byte
integer, whereas the priority on the RTCDataChannel object is an enum.
The integer has four recommended values in rtcweb-transport, but that
recommendation is only a "SHOULD", so technically any value from 0 to
65535 is possible. So something needs to define how these integer values
are mapped to the enum values.