Remove SSRCs from RTCRtpEncodingParameters.
Fixes #1174 and fixes #1435.

This was the decision at the June 28 virtual interim, to resolve the
differences between the WebRTC and ORTC definitions of the SSRC fields,
and to address the fact that the SSRC may change due to various reasons.

It was decided that there wasn't really a use for the SSRC fields in
WebRTC 1.0 in the first place, so we can just simplify things and
remove them.

Note that they *are* still present in stats, where they can be
correlated with an RTCRtpSender either by using the sender as a
selector, or by using `mediaTrackId` (assuming there aren't two senders
sending the same track).