Add a section summarizing different ICE candidate events.
Fixes #1213.

We've ended up with three different types of events that all use the
same "icecandidate" event, and are fired in different places. Which is,
understandably, confusing to most readers of the specification.

This PR adds a note to the RTCPeerConnectionIceEvent section summarizing
the different types of events, describing when they're fired, how
they're represented in the RTCPeerConnectionIceEvent, and what the
application should do with them.