Add ICE candidates only to the applicable descriptions.
Fixes #1227.

ICE candidates should only be added to descriptions that represent the
ICE generation of the candidate. For example, if
`currentLocalDescription` has ufrag "foo" and `pendingLocalDescription`
has ufrag "bar", because an ICE restart was initiated, then candidates
gathered for the old ICE generation ("foo") should be added only to
`currentLocalDescription`, and candidates gathered for the new ICE
generation should be added only to `pendingLocalDescription`.

Same goes for remote candidates, where the generation is inferred by
ufrag (or assumed, if ufrag isn't present).