Select sent codec via "codecPayloadType" field rather than reordering.
Fixes #1207.

Instead of selecting a codec by reordering the list from
`getParameters`, an application will now pick the payload type of
its preferred codec and set it in `RTCRtpEncodingParameters`. This
matches ORTC, and fixes the linked issue, because as long as the
selected codec is still present in a new offer/answer exchange, it will
stay selected.

If the selected codec is *not* present in a new offer/answer exchange,
the implementation will go back to its normal policy, which is the same
behavior as before this PR.

Note that this PR puts the `codecPayloadType` field on
`RTCRtpEncodingParameters`, which means the codec may be different
per-encoding, which wasn't possible before. If this is an issue, it
could be moved to the `RTCRtpParameters` level instead.