Describe how transport objects are assigned to senders/receivers.
As decided in the June virtual interim, DTLS/ICE transport objects will
only be created as a result of setting a local description. Also,
applying an answer may change the mapping between media descriptions and
transports (according to the rules in BUNDLE), which may result in
`sender.transport` (for example) pointing to a different object than
before.

This PR puts the actual responsibility for figuring out the media/
transport mapping on BUNDLE, using the "media transport" term, which in
this case is something that encapsulates the pair of ICE and DTLS
transports. Note that BUNDLE itself does not use that term (yet), but it
will, optimistically, if
https://github.com/cdh4u/draft-sdp-bundle/pull/19 is merged.

Note that this PR does not yet handle rollbacks, or RTCP mux
negotiation.