Adding relativeBitrate parameter to RTCRtpEncodingParameters.
Fixes #1625.

This provides more fine-grained control over the relative amount of bits
different encodings should be allocated. This allows the webrtc
implementation to respond quickly to changes in available capacity; the
alternative would be for an application to inspect the stats
periodically, and continually adjust `maxBitrate`, which would result in
wasted bandwidth and slower responsiveness to changing bandwidth
estimates (since polling stats and calling getParameters/setParameters
is not instantaneous).

This assumes that rtcweb-transports will update its definition of
"priority," such that it only controls the priority of streams in
network queues (via DSCP and SCTP priority values), and not the
relative amount of bitrate they're granted.