Fix setCodecPreferences note regarding offer/answer negotiation.
Fixes #1838.

The note said that, even if codec preferences are set, the list of
codecs in the answer will be a subset of those in the offer. Which JSEP
required before, but now does not:

  [T]he media formats in the answer MUST include at least one format
  that is present in the offer, but MAY include formats that are locally
  supported but not present in the offer.

So now, the subset rule is only a SHOULD in RFC3264, and not a MUST in
JSEP. So although most implementations are expected to honor the
"SHOULD", implementations can technically offer codecs in the answer
that were not in the offer. Which is now clarified in this note, with a
brief example.