Treat rejected m= sections (port of 0) as having "inactive" direction.
Fixes #1812.

This will:

- Cause `currentDirection` to change to `"inactive"`
- Cause the remote track to be removed from its stream(s), if
  applicable.

This seems pretty sensible, as the direction has no relevance for a
rejected m= section. No packets will be sent or received for a rejected
m= section (aka, stopped transceiver), so it's effecitvely `"inactive"`,
and it would be odd if a different direction in a rejected m= section
produced different behavior.

This also resolves a problem where remote tracks were possibly never
being removed from their streams, which would happen if you set a remote
description with a rejected m= section with "sendrecv" or "sendonly"
direction.