Adding note about potential consequences of stopping a transceiver.
Fixes #1858.

If a transceiver is stopped in between applying a remote offer and
creating an answer, and it's associated with the "offerer tagged" m=
section (the one listed first in "a=group:BUNDLE"), all other m=
sections in the BUNDLE group will be rejected as well.

The actual normative text should go in JSEP (see
https://github.com/rtcweb-wg/jsep/pull/849), but it's surprising enough
that it seems worth calling out in a note.