Adding note about the dangers of RTCDataChannelInit.negotiated.
Fixes #1879.

If one peer creates an out-of-band negotiated data channel and tries to
send with it before the other peer creates its corresponding data
channel, messages may be lost (even if the channel is reliable).

Raised this issue on the rtcweb mailing list here:
https://mailarchive.ietf.org/arch/msg/rtcweb/lIuiu91_L2nOh935eAqifrs_ius

But there's no clear consensus, with more people suggesting "do nothing"
than anything else. Given that this has always been an issue in Chrome
and Firefox and no one has really complained, maybe just adding this
note is sufficient.