Adding note about the relationship between tracks of senders/receivers. (#1911)
* Adding note about the relationship between tracks of senders/receivers.

Fixes #1169.

Clarifies that there's not a 1:1 relationship between tracks sent by
one PeerConnection and received by the other (due to `replaceTrack`),
and the IDs are not guaranteed to match. So it's better to think of
a 1:1 relationship between senders on one side and receivers on the
other side, rather than tracks.