Clarifying how the simulcast envelope is created.
Added text to articulate the difference between the offerer who
calls addTransceiver to define the simulcast envelope and the
answerer who is passive and has the envelope set up for them in a
call to setRemoteDescription.
Added steps to Set the RTCSessionDescription to indicate how the
the envelope should be set up or modified.

Fixes: #2014.