Clarify "multiple sources of media stitched together"
When `sender.replaceTrack` is executed, the sender's track is replaced but the `RTCRtpReceiver`still only sees a single RTP stream with a continuous sequence number space.  So in this case, the `RTCRtpReceiver` does not "stitch together" two distinct RTP streams, as would be required when receiving simulcast.  Therefore, the text is wrong and should be removed.

Fix for Issue https://github.com/w3c/webrtc-pc/issues/2171