Make RTCIceCandidatePair members required
In webrtc-pc this dictionary is only used as a return type of getSelectedCandidatePair(), which will include both local and remote. This is not a problem.

In https://w3c.github.io/webrtc-extensions/#rtcicetransport the dictionary is used as a non-optional argument, and Web IDL requires that dictionary arguments are optional if there are no required members. Fix this by making them required.