Convert RTCIceCandidatePair dictionary to an interface.
Addresses: #2930.

This commit changes the dictionary to an interface and required
dictionary members to readonly attributes backed by internal slots.

No constructor is described in the IDL, since RTCIceCandidatePair is
only created by the user agent as a result of the ICE agent forming an
ICE candidate pair. The application cannot construct arbitrary candidate
pairs.