Work around new requirements from respec for updated recs
instead of using a configuration key, respec now search for markup to determine whether a spec contains corrections or additions; but since the markup is added as post-processing in this spec, the detection fails. Instead, adding hidden markup to trigger the right effect.

That markup will have to have the classes proposed-addition / proposed-correction if/when we reach that status